+++ R/find_parameters_other.R
@@ -321,3 +321,67 @@ find_parameters.oohbchoice <- function(x, flatten = FALSE, ...) {
     out
   }
 }
+
+
+#' @export
+find_parameters.lcmm <- function(x, component = "all", flatten = FALSE, ...) {
+  coefficients <- stats::coef(x)
+  params <- names(coefficients)[!startsWith(names(coefficients), "cholesky ")]
+  component <- validate_argument(
+    component,
+    c("all", "membership", "longitudinal", "conditional", "beta", "splines", "linear")
+  )
+
+  n_membership <- x$N[1]
+  n_longitudinal <- x$N[2]
+
+  if (x$linktype == 1) {
+    type <- "Beta"
+  } else if (x$linktype == 2) {
+    type <- "I-splines"
+  } else if (x$linktype == 3) {
+    ## TODO: check if this is correct
+  } else {
+    type <- "Linear"
+  }
+
+  out <- list(
+    membership = params[!startsWith(params, type)][seq_len(n_membership)],
+    longitudinal = params[!startsWith(params, type)][
+      (n_membership + 1):(n_membership + n_longitudinal)
+    ],
+    extra = params[startsWith(params, type)]
+  )
+  names(out)[3] <- switch(
+    type,
+    "Beta" = "beta",
+    "I-splines" = "splines",
+    "Linear" = "linear"
+  )
+
+  out <- compact_list(out)
+
+  .filter_parameters(
+    out,
+    effects = "all",
+    component = component,
+    flatten = flatten,
+    recursive = FALSE
+  )
+}
+
+
+#' @export
+find_parameters.externX <- function(x, flatten = FALSE, ...) {
+  coefficients <- stats::coef(x)
+  out <- list(conditional = names(coefficients))
+
+  if (flatten) {
+    unique(unlist(out, use.names = FALSE))
+  } else {
+    out
+  }
+}
+
+#' @export
+find_parameters.externVar <- find_parameters.externX