+++ R/get_loglikelihood.R
@@ -364,6 +364,17 @@ get_loglikelihood.phylolm <- function(x, check_response = FALSE, verbose = TRUE,
 #' @export
 get_loglikelihood.phyloglm <- get_loglikelihood.phylolm
 
+#' @export
+get_loglikelihood.lcmm <- function(x, verbose = TRUE, ...) {
+  x$loglik
+}
+
+#' @export
+get_loglikelihood.externX <- get_loglikelihood.lcmm
+
+#' @export
+get_loglikelihood.externVar <- get_loglikelihood.lcmm
+
 
 # Methods WITH individual LLs ---------------------------------------------